requirements.txt
absl-py==1.3.0
altair==4.2.0
astunparse==1.6.3
attrs==22.1.0
autopep8==2.0.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.2.0
catboost==1.1.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
flatbuffers==22.11.23
fonttools==4.38.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
lightgbm==3.3.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
MouseInfo==0.1.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
plotly==5.11.0
protobuf==3.19.6
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pydeck==0.8.0
PyGetWindow==0.0.9
Pygments==2.13.0
Pympler==1.0.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.2
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
six==1.16.0
smmap==5.0.0
stopwords==1.0.0
streamlit==1.15.1
tenacity==8.1.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==2.1.1
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
typing-extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.7.1
zipp==3.11.0
